Backlog: riordino drag&drop e stato derivato da Jira
- priority_order da Integer a Float: permette di inserire una riga "a
  meta'" tra due esistenti via drag&drop, senza rinumerare tutto il
  backlog (l'ordine manuale resta comunque preservato dopo la sync,
  come gia' accadeva prima di questa modifica)
- BacklogPage: nuova maniglia di trascinamento per riordinare le righe
  (drag&drop nativo HTML5, nessuna dipendenza aggiuntiva); al rilascio
  la nuova posizione e' la media tra i due vicini nella lista visibile
- BacklogItem.status ora deriva da jira_status quando presente (Done/
  In Progress/altro->To Do), invece del solo calcolo locale da
  actual_start/actual_finish che poteva disallinearsi da Jira; il
  fallback sulle date resta per gli item creati a mano senza sync

+++ backend/app/schemas.py
@@ -59,7 +59,7 @@ class BudgetLine(BudgetLineBase):
 
 class BacklogItemBase(BaseModel):
     jira_key: str
-    priority_order: int = 0
+    priority_order: float = 0
     ready_for_refinement: bool = False
     in_scope: bool = True
     planned: bool = False
@@ -80,7 +80,7 @@ class BacklogItemCreate(BacklogItemBase):
 
 
 class BacklogItemUpdate(BaseModel):
-    priority_order: int | None = None
+    priority_order: float | None = None
     ready_for_refinement: bool | None = None
     in_scope: bool | None = None
     planned: bool | None = None
